gpu_unai: 64-bit fixes from Notaz pcsx_rearmed

+++ src/gpu/gpu_unai/gpu.cpp
@@ -465,7 +465,9 @@ void  GPU_readDataMem(u32* dmaAddress, s32 dmaCount)
 	{
 		if ((&pvram[px])>(VIDEO_END)) pvram-=512*1024;
 		// lower 16 bit
-		u32 data = (unsigned long)pvram[px];
+		//senquack - 64-bit fix (from notaz)
+		//u32 data = (unsigned long)pvram[px];
+		u32 data = (u32)pvram[px];
 
 		if (++px>=x_end) 
 		{
@@ -475,7 +477,9 @@ void  GPU_readDataMem(u32* dmaAddress, s32 dmaCount)
 
 		if ((&pvram[px])>(VIDEO_END)) pvram-=512*1024;
 		// higher 16 bit (always, even if it's an odd width)
-		data |= (unsigned long)(pvram[px])<<16;
+		//senquack - 64-bit fix (from notaz)
+		//data |= (unsigned long)(pvram[px])<<16;
+		data |= (u32)(pvram[px])<<16;
 		
 		*dmaAddress++ = data;
